Add Settings Page and Remove SettingsModal
